Don't import from Data.Typeable.Internal when possible (i.e., add support for GHC 8.2)
GHC 8.2 reshuffled around the internals of `Typeable` quite a bit, and as a
result, `base` no longer exports the `Data.Typeable.Internal` module in that
version. This causes `hspec-core` (and `hspec-meta`) to fail to build on GHC
8.2.

But this is quite easily fixable, as `hspec` is only using
`Data.Typeable.Internal` to import `tyConModule` and `tyConName`, which have
been exported from the public `Data.Typeable` interface since GHC 7.4, which
is the minimum version of GHC that `hspec` supports.

While this fixes the `hspec-core` issue, I don't know how to make the same
change to `hspec-meta`, as I don't know where it lives.