Asynchronous formatting (optional)
Hspec formatters only receive events about the first test being run, even
if there are multiple tests running in parallel.

When the new option is enabled, formatters receive events about all the tests
running. The options is not exposed by a flag, it can only be enabled
programmatically. Currently this happens implicitly with the choice of formatter.