New formatter "specdyn"
Designed as an extension of "specdoc", "specdyn" is able to show
the example(s) that is being executed, which is convenient with
slow / long running tests combined with parallel.

This change introduces a new effect TextF for describing computations
that append and update text blocks on the terminal. The existing
formatters are refactored on top of the new effect.