Call release only when acquired in hooks
This fixes an issue when the `release` action in `aroundAll_` and `aroundAllWith` is called when the `acquire` action was not, causing a deadlock. This situation could happen with multiple `aroundAll_` actions when one of the earlier `acquire` action has failed but the `release` actions of uncalled `acquire` still get executed.