Prevent deadlocks on unguarded exceptions in runner
The runner operates under the assumption that all exceptions from spec
items are handled and converted to `Result`s.  This is ensured by
`safeEvaluateExample`.

If a user violates this assumption by manually modifying a spec item
with `mapSpecItem_` and omitting proper exception handling then this can
lead to a deadlock.

This change addresses this by no longer relying on that assumption.