Support passing --format more than once
We use a JUnit formatter that produces an XML file for external
analysis. When we do this, we'd like to still output the normal format
in the build logs, since viewing them there is also useful. Many of our
Engineers get confused when their build fails tests and they navigate to
the GitHub action and see no output. The summary and annotations
produced from our JUnit report are both less intuitive to find (although
more useful once they are found, which is why we do it).

This commit changes the behavior when passing `--format` multiple times
from "last wins" to "use them all". That way, we can pass `--format
checks --format junit` to achieve what we want.