Add --add-format option
This option can be used when you want to use a formatter in addition to
the default. Without this option, one would have to use `--format checks
--format junit` to get that behavior, which duplicates knowledge of the
default formatter and can easily be missed. Now, they can use
`--add-format junit` to accomplish it more directly.

The flag and help are placeholder; I'm happy to name it and document it
however you like.